# openapi.yaml
openapi: 3.0.0
info:
  description: "Nncof_EventsSubscription Service API.  \n© 2025, 3GPP Organizational\
    \ Partners (ARIB, ATIS, CCSA, ETSI, TSDSI, TTA, TTC).  \nAll rights reserved.\n"
  title: Nncof_EventsSubscription
  version: 1.0.0
...
servers:
- description: SwaggerHub API Auto Mocking
  url: https://virtserver.swaggerhub.com/ETRI_INRS_TEAM/NCOF_Nnwdaf_EventSubscription/1.0.0
- description: SwaggerHub API Auto Mocking
  url: https://virtserver.swaggerhub.com/ETRI_INRS_TEAM/NCOF_Nncof_EventSubscription/1.0.0
- url: "{apiRoot}/nnwdaf-eventssubscription/v1"
  variables:
    apiRoot:
      default: https://example.com
      description: apiRoot as defined in clause 4.4 of 3GPP TS 29.501.
security:
- {}
- oAuth2ClientCredentials:
  - nncof-eventssubscription
paths:
...
  /subscriptions/{subscriptionId}:
    delete:
      operationId: DeleteNCOFEventsSubscription
      parameters:
      - description: String identifying a subscription to the Nncof_EventsSubscription
          Service
        explode: false
        in: path
        name: subscriptionId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: |
            No Content. The Individual NCOF Event Subscription resource matching the subscriptionId was deleted.
        "307":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedirectResponse'
          description: Temporary Redirect
          headers:
            Location:
              description: The URI pointing to the resource located on the redirect
                target
              required: true
              schema:
                type: string
            "3gpp-Sbi-Target-Nf-Id":
              description: |
                'Identifier of target NF (service) instance towards which the request is redirected'
              schema:
                type: string
        "308":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedirectResponse'
          description: Permanent Redirect
          headers:
            Location:
              description: The URI pointing to the resource located on the redirect
                target
              required: true
              schema:
                type: string
            "3gpp-Sbi-Target-Nf-Id":
              description: |
                'Identifier of target NF (service) instance towards which the request is redirected'
              schema:
                type: string
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Bad request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Unauthorized
        "403":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Forbidden
        "404":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Not Found
        "429":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Too Many Requests
        "500":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Internal Server Error
        "501":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Not Implemented
        "502":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Bad Gateway
        "503":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Service Unavailable
        default:
          description: Generic Error
      summary: Delete an existing Individual NCOF Events Subscription
      tags:
      - Individual NCOF Events Subscription
    put:
      operationId: UpdateNCOFEventsSubscription
      parameters:
      - description: String identifying a subscription to the Nncof_EventsSubscription
          Service.
        explode: false
        in: path
        name: subscriptionId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NncofEventsSubscription'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NncofEventsSubscription'
          description: |
            The Individual NCOF Event Subscription resource was modified successfully and a representation of that resource is returned.
        "204":
          description: The Individual NCOF Event Subscription resource was modified
            successfully.
        "307":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedirectResponse'
          description: Temporary Redirect
          headers:
            Location:
              description: The URI pointing to the resource located on the redirect
                target
              required: true
              schema:
                type: string
            "3gpp-Sbi-Target-Nf-Id":
              description: |
                'Identifier of target NF (service) instance towards which the request is redirected'
              schema:
                type: string
        "308":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedirectResponse'
          description: Permanent Redirect
          headers:
            Location:
              description: The URI pointing to the resource located on the redirect
                target
              required: true
              schema:
                type: string
            "3gpp-Sbi-Target-Nf-Id":
              description: |
                'Identifier of target NF (service) instance towards which the request is redirected'
              schema:
                type: string
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Bad request
        "401":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Unauthorized
        "403":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Forbidden
        "404":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Not Found
        "411":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Length Required
        "413":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Content Too Large
        "415":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Unsupported Media Type
        "429":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Too Many Requests
        "500":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Internal Server Error
        "501":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Not Implemented
        "502":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Bad Gateway
        "503":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: Service Unavailable
        default:
          description: Generic Error
      summary: Update an existing Individual NCOF Events Subscription
      tags:
      - Individual NCOF Events Subscription
...
components:
...
  schemas:
    NncofEventsSubscription:
      description: Represents an Individual NCOF Event Subscription resource.
      example:
        supportedFeatures: supportedFeatures
        notificationURI: notificationURI
        prevSub: prevSub
        notifCorrId: notifCorrId
        eventNotifications:
        - nfLoadLevelInfos:
          - nfInstanceId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            snssai:
              sd: sd
              sst: 37
            nfStorageUsage: 3
            nfLoadLevelpeak: 1
            nfStatus:
              statusRegistered: 93
              statusUndiscoverable: 87
              statusUnregistered: 67
            nfCpuUsage: 9
            confidence: 0
            nfType: null
            nfSetId: nfSetId
            nfMemoryUsage: 6
            nfLoadLevelAverage: 6
            nfLoadAvgInAoi: 2
          - nfInstanceId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            snssai:
              sd: sd
              sst: 37
            nfStorageUsage: 3
            nfLoadLevelpeak: 1
            nfStatus:
              statusRegistered: 93
              statusUndiscoverable: 87
              statusUnregistered: 67
            nfCpuUsage: 9
            confidence: 0
            nfType: null
            nfSetId: nfSetId
            nfMemoryUsage: 6
            nfLoadLevelAverage: 6
            nfLoadAvgInAoi: 2
        - nfLoadLevelInfos:
          - nfInstanceId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            snssai:
              sd: sd
              sst: 37
            nfStorageUsage: 3
            nfLoadLevelpeak: 1
            nfStatus:
              statusRegistered: 93
              statusUndiscoverable: 87
              statusUnregistered: 67
            nfCpuUsage: 9
            confidence: 0
            nfType: null
            nfSetId: nfSetId
            nfMemoryUsage: 6
            nfLoadLevelAverage: 6
            nfLoadAvgInAoi: 2
          - nfInstanceId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
            snssai:
              sd: sd
              sst: 37
            nfStorageUsage: 3
            nfLoadLevelpeak: 1
            nfStatus:
              statusRegistered: 93
              statusUndiscoverable: 87
              statusUnregistered: 67
            nfCpuUsage: 9
            confidence: 0
            nfType: null
            nfSetId: nfSetId
            nfMemoryUsage: 6
            nfLoadLevelAverage: 6
            nfLoadAvgInAoi: 2
        consNfInfo: consNfInfo
        eventSubscriptions:
        - extraReportReq:
            endTs: 2000-01-23T04:56:07.000+00:00
            startTs: 2000-01-23T04:56:07.000+00:00
          tgtUe:
            supis:
            - null
            - null
            gpsis:
            - null
            - null
            anyUe: true
            intGroupIds:
            - null
            - null
          nfLoadLvlThds:
          - svcExpLevel: 1.2315135
            nfStorageUsage: 5
            minTrafficRate: minTrafficRate
            avgTrafficRate: avgTrafficRate
            varPacketDelay: 3.6160767
            nfMemoryUsage: 5
            speed: 1.0246457
            avgPacketDelay: 1
            varPacketLossRate: 7.386282
            varTrafficRate: 2.302136
            congLevel: 0
            nfCpuUsage: 1
            maxTrafficRate: maxTrafficRate
            maxPacketDelay: 1
            nfLoadLevel: 6
            aggTrafficRate: aggTrafficRate
            avgPacketLossRate: 202
            maxPacketLossRate: 414
          - svcExpLevel: 1.2315135
            nfStorageUsage: 5
            minTrafficRate: minTrafficRate
            avgTrafficRate: avgTrafficRate
            varPacketDelay: 3.6160767
            nfMemoryUsage: 5
            speed: 1.0246457
            avgPacketDelay: 1
            varPacketLossRate: 7.386282
            varTrafficRate: 2.302136
            congLevel: 0
            nfCpuUsage: 1
            maxTrafficRate: maxTrafficRate
            maxPacketDelay: 1
            nfLoadLevel: 6
            aggTrafficRate: aggTrafficRate
            avgPacketLossRate: 202
            maxPacketLossRate: 414
          nfSetIds:
          - null
          - null
          snssaia:
          - sd: sd
            sst: 37
          - sd: sd
            sst: 37
          appIds:
          - null
          - null
          nfTypes:
          - NRF
          - NRF
          evtReq:
            notifMethod: null
            partitionCriteria:
            - TAC
            - TAC
            grpRepTime: 4
            notifFlag: ACTIVATE
            mutingSetting:
              maxNoOfNotif: 5
              durationBufferedNotif: 9
            monDur: 2000-01-23T04:56:07.000+00:00
            immRep: true
            maxReportNbr: 0
            repPeriod: 7
            sampRatio: 12
            notifFlagInstruct:
              bufferedNotifs: SEND_ALL
              subscription: CLOSE
          congThresholds:
          - svcExpLevel: 1.2315135
            nfStorageUsage: 5
            minTrafficRate: minTrafficRate
            avgTrafficRate: avgTrafficRate
            varPacketDelay: 3.6160767
            nfMemoryUsage: 5
            speed: 1.0246457
            avgPacketDelay: 1
            varPacketLossRate: 7.386282
            varTrafficRate: 2.302136
            congLevel: 0
            nfCpuUsage: 1
            maxTrafficRate: maxTrafficRate
            maxPacketDelay: 1
            nfLoadLevel: 6
            aggTrafficRate: aggTrafficRate
            avgPacketLossRate: 202
            maxPacketLossRate: 414
          - svcExpLevel: 1.2315135
            nfStorageUsage: 5
            minTrafficRate: minTrafficRate
            avgTrafficRate: avgTrafficRate
            varPacketDelay: 3.6160767
            nfMemoryUsage: 5
            speed: 1.0246457
            avgPacketDelay: 1
            varPacketLossRate: 7.386282
            varTrafficRate: 2.302136
            congLevel: 0
            nfCpuUsage: 1
            maxTrafficRate: maxTrafficRate
            maxPacketDelay: 1
            nfLoadLevel: 6
            aggTrafficRate: aggTrafficRate
            avgPacketLossRate: 202
            maxPacketLossRate: 414
          anySlice: true
          notificationMethod: PERIODIC
          event: SLICE_LOAD_LEVEL
          exptAnaType: MOBILITY
          nfInstanceIds:
          - null
          - null
        - extraReportReq:
            endTs: 2000-01-23T04:56:07.000+00:00
            startTs: 2000-01-23T04:56:07.000+00:00
          tgtUe:
            supis:
            - null
            - null
            gpsis:
            - null
            - null
            anyUe: true
            intGroupIds:
            - null
            - null
          nfLoadLvlThds:
          - svcExpLevel: 1.2315135
            nfStorageUsage: 5
            minTrafficRate: minTrafficRate
            avgTrafficRate: avgTrafficRate
            varPacketDelay: 3.6160767
            nfMemoryUsage: 5
            speed: 1.0246457
            avgPacketDelay: 1
            varPacketLossRate: 7.386282
            varTrafficRate: 2.302136
            congLevel: 0
            nfCpuUsage: 1
            maxTrafficRate: maxTrafficRate
            maxPacketDelay: 1
            nfLoadLevel: 6
            aggTrafficRate: aggTrafficRate
            avgPacketLossRate: 202
            maxPacketLossRate: 414
          - svcExpLevel: 1.2315135
            nfStorageUsage: 5
            minTrafficRate: minTrafficRate
            avgTrafficRate: avgTrafficRate
            varPacketDelay: 3.6160767
            nfMemoryUsage: 5
            speed: 1.0246457
            avgPacketDelay: 1
            varPacketLossRate: 7.386282
            varTrafficRate: 2.302136
            congLevel: 0
            nfCpuUsage: 1
            maxTrafficRate: maxTrafficRate
            maxPacketDelay: 1
            nfLoadLevel: 6
            aggTrafficRate: aggTrafficRate
            avgPacketLossRate: 202
            maxPacketLossRate: 414
          nfSetIds:
          - null
          - null
          snssaia:
          - sd: sd
            sst: 37
          - sd: sd
            sst: 37
          appIds:
          - null
          - null
          nfTypes:
          - NRF
          - NRF
          evtReq:
            notifMethod: null
            partitionCriteria:
            - TAC
            - TAC
            grpRepTime: 4
            notifFlag: ACTIVATE
            mutingSetting:
              maxNoOfNotif: 5
              durationBufferedNotif: 9
            monDur: 2000-01-23T04:56:07.000+00:00
            immRep: true
            maxReportNbr: 0
            repPeriod: 7
            sampRatio: 12
            notifFlagInstruct:
              bufferedNotifs: SEND_ALL
              subscription: CLOSE
          congThresholds:
          - svcExpLevel: 1.2315135
            nfStorageUsage: 5
            minTrafficRate: minTrafficRate
            avgTrafficRate: avgTrafficRate
            varPacketDelay: 3.6160767
            nfMemoryUsage: 5
            speed: 1.0246457
            avgPacketDelay: 1
            varPacketLossRate: 7.386282
            varTrafficRate: 2.302136
            congLevel: 0
            nfCpuUsage: 1
            maxTrafficRate: maxTrafficRate
            maxPacketDelay: 1
            nfLoadLevel: 6
            aggTrafficRate: aggTrafficRate
            avgPacketLossRate: 202
            maxPacketLossRate: 414
          - svcExpLevel: 1.2315135
            nfStorageUsage: 5
            minTrafficRate: minTrafficRate
            avgTrafficRate: avgTrafficRate
            varPacketDelay: 3.6160767
            nfMemoryUsage: 5
            speed: 1.0246457
            avgPacketDelay: 1
            varPacketLossRate: 7.386282
            varTrafficRate: 2.302136
            congLevel: 0
            nfCpuUsage: 1
            maxTrafficRate: maxTrafficRate
            maxPacketDelay: 1
            nfLoadLevel: 6
            aggTrafficRate: aggTrafficRate
            avgPacketLossRate: 202
            maxPacketLossRate: 414
          anySlice: true
          notificationMethod: PERIODIC
          event: SLICE_LOAD_LEVEL
          exptAnaType: MOBILITY
          nfInstanceIds:
          - null
          - null
        failEventReports:
        - failEventReports
        - failEventReports
      properties:
        eventSubscriptions:
          description: Subscribed events
          items:
            $ref: '#/components/schemas/EventSubscription'
          minItems: 1
          title: eventSubscriptions
          type: array
        notificationURI:
          description: String providing an URI formatted according to RFC 3986.
          title: type
          type: string
        notifCorrId:
          description: Notification correlation identifier.
          title: notifCorrId
          type: string
        supportedFeatures:
          description: |
            A string used to indicate the features supported by an API that is used as defined in clause  6.6 in 3GPP TS 29.500. The string shall contain a bitmask indicating supported features in  hexadecimal representation Each character in the string shall take a value of "0" to "9",  "a" to "f" or "A" to "F" and shall represent the support of 4 features as described in  table 5.2.2-3. The most significant character representing the highest-numbered features shall  appear first in the string, and the character representing features 1 to 4 shall appear last  in the string. The list of features and their numbering (starting with 1) are defined  separately for each API. If the string contains a lower number of characters than there are  defined features for an API, all features that would be represented by characters that are not  present in the string are not supported.
          pattern: "^[A-Fa-f0-9]*$"
          title: supportedFeatures
          type: string
        eventNotifications:
          items:
            $ref: '#/components/schemas/EventNotification'
          minItems: 0
          title: eventNotifications
          type: array
        failEventReports:
          items:
            type: string
          minItems: 1
          title: failEventReports
          type: array
        prevSub:
          title: prevSub
          type: string
        consNfInfo:
          title: consNfInfo
          type: string
      required:
      - eventSubscriptions
      title: NncofEventsSubscription
      type: object
...
    EventNotification:
      description: Represents a notification on events that occurred.
      example:
        nfLoadLevelInfos:
        - nfInstanceId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          snssai:
            sd: sd
            sst: 37
          nfStorageUsage: 3
          nfLoadLevelpeak: 1
          nfStatus:
            statusRegistered: 93
            statusUndiscoverable: 87
            statusUnregistered: 67
          nfCpuUsage: 9
          confidence: 0
          nfType: null
          nfSetId: nfSetId
          nfMemoryUsage: 6
          nfLoadLevelAverage: 6
          nfLoadAvgInAoi: 2
        - nfInstanceId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          snssai:
            sd: sd
            sst: 37
          nfStorageUsage: 3
          nfLoadLevelpeak: 1
          nfStatus:
            statusRegistered: 93
            statusUndiscoverable: 87
            statusUnregistered: 67
          nfCpuUsage: 9
          confidence: 0
          nfType: null
          nfSetId: nfSetId
          nfMemoryUsage: 6
          nfLoadLevelAverage: 6
          nfLoadAvgInAoi: 2
      properties:
        nfLoadLevelInfos:
          items:
            $ref: '#/components/schemas/NfLoadLevelInformation'
          title: nfLoadLevelInfos
          type: array
      title: EventNotification
      type: object
...
    ProblemDetails:
      description: Provides additional information in an error response.
      properties:
        type:
          description: String providing an URI formatted according to RFC 3986.
          title: type
          type: string
        title:
          title: title
          type: string
        status:
          title: status
          type: integer
        detail:
          description: A human-readable explanation specific to this occurrence of
            the problem.
          title: detail
          type: string
        instance:
          description: String providing an URI formatted according to RFC 3986.
          title: type
          type: string
        cause:
          description: |
            A machine-readable application error cause specific to this occurrence of the problem.  This IE should be present and provide application-related error information, if available.
          title: cause
          type: string
        invalidParams:
          items:
            $ref: '#/components/schemas/InvalidParam'
          minItems: 1
          title: invalidParams
          type: array
        supportedFeatures:
          description: |
            A string used to indicate the features supported by an API that is used as defined in clause  6.6 in 3GPP TS 29.500. The string shall contain a bitmask indicating supported features in  hexadecimal representation Each character in the string shall take a value of "0" to "9",  "a" to "f" or "A" to "F" and shall represent the support of 4 features as described in  table 5.2.2-3. The most significant character representing the highest-numbered features shall  appear first in the string, and the character representing features 1 to 4 shall appear last  in the string. The list of features and their numbering (starting with 1) are defined  separately for each API. If the string contains a lower number of characters than there are  defined features for an API, all features that would be represented by characters that are not  present in the string are not supported.
          pattern: "^[A-Fa-f0-9]*$"
          title: supportedFeatures
          type: string
        supportedApiVersions:
          items:
            type: string
          minItems: 1
          title: supportedApiVersions
          type: array
      title: ProblemDetails
      type: object
...
    InvalidParam:
      description: It contains an invalid parameter and a related description.
      properties:
        param:
          description: |
            If the invalid parameter is an attribute in a JSON body, this IE shall contain the  attribute's name and shall be encoded as a JSON Pointer. If the invalid parameter is  an HTTP header, this IE shall be formatted as the concatenation of the string "header "  plus the name of such header. If the invalid parameter is a query parameter, this IE  shall be formatted as the concatenation of the string "query " plus the name of such  query parameter. If the invalid parameter is a variable part in the path of a resource  URI, this IE shall contain the name of the variable, including the symbols "{" and "}"  used in OpenAPI specification as the notation to represent variable path segments.
          title: param
          type: string
        reason:
          description: |
            A human-readable reason, e.g. "must be a positive integer". In cases involving failed  operations in a PATCH request, the reason string should identify the operation that  failed using the operation's array index to assist in correlation of the invalid  parameter with the failed operation, e.g." Replacement value invalid for attribute  (failed operation index= 4)"
          title: reason
          type: string
      required:
      - param
      title: InvalidParam
      type: object
...
    RedirectResponse:
      description: |
        The response shall include a Location header field containing a different URI  (pointing to a different URI of an other service instance), or the same URI if a request  is redirected to the same target resource via a different SCP.
      example:
        targetSepp: targetSepp
        cause: cause
        targetScp: targetScp
      properties:
        cause:
          title: cause
          type: string
        targetScp:
          description: String providing an URI formatted according to RFC 3986.
          title: type
          type: string
        targetSepp:
          description: String providing an URI formatted according to RFC 3986.
          title: type
          type: string
      title: RedirectResponse
      type: object
...
  securitySchemes:
    oAuth2ClientCredentials:
      flows:
        clientCredentials:
          scopes:
            nncof-eventssubscription: Access to the Nncof_EventsSubscription API
            nncof-eventssubscription:transfer: |
              Access to service operations applying to NCOF event subscription transfer.
          tokenUrl: "{nrfApiRoot}/oauth2/token"
      type: oauth2